+++ voile-util/src/lib.rs
@@ -7,6 +7,7 @@ to help the development of other dependently-typed lambda calculus type checkers
 It contains helper functions for the [Pest] parser
 (supported via optional feature `parser`),
 a non-empty vector, some location utils, a unique-ID type with utils,
+a lisp parser for term generation,
 and universe level utilities (with omega).
 
 All dependencies are optional, thus very lightweight.